+++ client/src/cmdlfjablotron.c
@@ -152,7 +152,8 @@ static int CmdJablotronClone(const char *Cmd) {
 
     CLIParserContext *ctx;
     CLIParserInit(&ctx, "lf jablotron clone",
-                  "clone a Jablotron tag to a T55x7, Q5/T5555 or EM4305/4469 tag.",
+                  "clone a Jablotron tag to a T55x7, Q5/T5555 or EM4305/4469 tag.\n"
+                  "Tag must be on the antenna when issuing this command.",
                   "lf jablotron clone --cn 01b669\n"
                   "lf jablotron clone --q5 --cn 01b669 -> encode for Q5/T5555 tag\n"
                   "lf jablotron clone --em --cn 01b669 -> encode for EM4305/4469"